Pass project status programmatically to react render.
Beef up the above the fold area so below the fold makes visual sense.
Use Lorem-ipsum for the above-the-fold content because it doesn't matter. Just filler.
Configure node-jsx to use .jsx extension so that it matches the web pack rules.  Finally rendering html on the server.
Render the below-the fold content using react on the server.
Pass the markup from react into the els below the fold template.
Create a below react component so we have a single, common point of entry for rendering on the client and the server.
Wrap the client-side js scripts with a flag so you can turn them off for demoing.